@@ -1,12 +1,11 @@
 name: geofencing
 description: A geofencing plugin for Flutter applications.
 version: 0.1.0
-author: Ben Konyi <bkonyi@google.com>
 homepage: https://github.com/flutter/plugins
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: "^1.12.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: "^2.0.0"
 
 dependencies:
   flutter:
