@@ -9,6 +9,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 
 import 'package:geofencing/geofencing.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 void main() => runApp(MyApp());
 
@@ -20,28 +21,30 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   String geofenceState = 'N/A';
   List<String> registeredGeofences = [];
-  double latitude = 37.419851;
-  double longitude = -122.078818;
-  double radius = 150.0;
+  double latitude = 50.00187;
+  double longitude = 36.23866;
+  double radius = 200.0;
   ReceivePort port = ReceivePort();
   final List<GeofenceEvent> triggers = <GeofenceEvent>[
     GeofenceEvent.enter,
-    GeofenceEvent.dwell,
     GeofenceEvent.exit
   ];
   final AndroidGeofencingSettings androidSettings = AndroidGeofencingSettings(
-      initialTrigger: <GeofenceEvent>[
-        GeofenceEvent.enter,
-        GeofenceEvent.exit,
-        GeofenceEvent.dwell
-      ],
-      loiteringDelay: 1000 * 60);
+    initialTrigger: <GeofenceEvent>[
+      GeofenceEvent.enter,
+      GeofenceEvent.exit,
+    ],
+    loiteringDelay: 0,
+    notificationResponsiveness: 0,
+  );
 
   @override
   void initState() {
     super.initState();
     IsolateNameServer.registerPortWithName(
-        port.sendPort, 'geofencing_send_port');
+      port.sendPort,
+      'geofencing_send_port',
+    );
     port.listen((dynamic data) {
       print('Event: $data');
       setState(() {
@@ -51,14 +54,44 @@ class _MyAppState extends State<MyApp> {
     initPlatformState();
   }
 
+  void registerGeofence() async {
+    final firstPermission = await Permission.locationWhenInUse.request();
+    final secondPermission = await Permission.locationAlways.request();
+    if (firstPermission.isGranted && secondPermission.isGranted) {
+      await GeofencingManager.registerGeofence(
+        GeofenceRegion(
+          'mtv',
+          latitude,
+          longitude,
+          radius,
+          triggers,
+          androidSettings,
+        ),
+        callback,
+      );
+      final registeredIds = await GeofencingManager.getRegisteredGeofenceIds();
+      setState(() {
+        registeredGeofences = registeredIds;
+      });
+    }
+  }
+
+  void unregisteGeofence() async {
+    await GeofencingManager.removeGeofenceById('mtv');
+    final registeredIds = await GeofencingManager.getRegisteredGeofenceIds();
+    setState(() {
+      registeredGeofences = registeredIds;
+    });
+  }
+
+  @pragma('vm:entry-point')
   static void callback(List<String> ids, Location l, GeofenceEvent e) async {
     print('Fences: $ids Location $l Event: $e');
     final SendPort send =
-    IsolateNameServer.lookupPortByName('geofencing_send_port');
+        IsolateNameServer.lookupPortByName('geofencing_send_port');
     send?.send(e.toString());
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> initPlatformState() async {
     print('Initializing...');
     await GeofencingManager.initialize();
@@ -90,44 +123,16 @@ class _MyAppState extends State<MyApp> {
                   children: <Widget>[
                     Text('Current state: $geofenceState'),
                     Center(
-                      child: RaisedButton(
+                      child: TextButton(
                         child: const Text('Register'),
-                        onPressed: () {
-                          if (latitude == null) {
-                            setState(() => latitude = 0.0);
-                          }
-                          if (longitude == null) {
-                            setState(() => longitude = 0.0);
-                          }
-                          if (radius == null) {
-                            setState(() => radius = 0.0);
-                          }
-                          GeofencingManager.registerGeofence(
-                              GeofenceRegion(
-                                  'mtv', latitude, longitude, radius, triggers,
-                                  androidSettings: androidSettings),
-                              callback).then((_) {
-                            GeofencingManager.getRegisteredGeofenceIds().then((value) {
-                              setState(() {
-                                registeredGeofences = value;
-                              });
-                            });
-                          });
-                        },
+                        onPressed: registerGeofence,
                       ),
                     ),
                     Text('Registered Geofences: $registeredGeofences'),
                     Center(
-                      child: RaisedButton(
+                      child: TextButton(
                         child: const Text('Unregister'),
-                        onPressed: () =>
-                            GeofencingManager.removeGeofenceById('mtv').then((_) {
-                              GeofencingManager.getRegisteredGeofenceIds().then((value){
-                                setState(() {
-                                  registeredGeofences = value;
-                                });
-                              });
-                            }),
+                        onPressed: unregisteGeofence,
                       ),
                     ),
                     TextField(
@@ -136,17 +141,17 @@ class _MyAppState extends State<MyApp> {
                       ),
                       keyboardType: TextInputType.number,
                       controller:
-                      TextEditingController(text: latitude.toString()),
+                          TextEditingController(text: latitude.toString()),
                       onChanged: (String s) {
                         latitude = double.tryParse(s);
                       },
                     ),
                     TextField(
                         decoration:
-                        const InputDecoration(hintText: 'Longitude'),
+                            const InputDecoration(hintText: 'Longitude'),
                         keyboardType: TextInputType.number,
                         controller:
-                        TextEditingController(text: longitude.toString()),
+                            TextEditingController(text: longitude.toString()),
                         onChanged: (String s) {
                           longitude = double.tryParse(s);
                         }),
@@ -154,7 +159,7 @@ class _MyAppState extends State<MyApp> {
                         decoration: const InputDecoration(hintText: 'Radius'),
                         keyboardType: TextInputType.number,
                         controller:
-                        TextEditingController(text: radius.toString()),
+                            TextEditingController(text: radius.toString()),
                         onChanged: (String s) {
                           radius = double.tryParse(s);
                         }),
