@@ -72,10 +72,13 @@ class GeofenceRegion {
   final AndroidGeofencingSettings androidSettings;
 
   GeofenceRegion(
-      this.id, double latitude, double longitude, this.radius, this.triggers,
-      {AndroidGeofencingSettings androidSettings})
-      : location = Location(latitude, longitude),
-        androidSettings = (androidSettings ?? AndroidGeofencingSettings());
+    this.id,
+    double latitude,
+    double longitude,
+    this.radius,
+    this.triggers,
+    this.androidSettings,
+  ) : location = Location(latitude, longitude);
 
   List<dynamic> _toArgs() {
     final int triggerMask = triggers.fold(
@@ -102,10 +105,12 @@ class GeofencingManager {
 
   /// Initialize the plugin and request relevant permissions from the user.
   static Future<void> initialize() async {
-    final CallbackHandle callback =
+    final CallbackHandle? callback =
         PluginUtilities.getCallbackHandle(callbackDispatcher);
-    await _channel.invokeMethod('GeofencingPlugin.initializeService',
-        <dynamic>[callback.toRawHandle()]);
+    if (callback != null) {
+      await _channel.invokeMethod('GeofencingPlugin.initializeService',
+          <dynamic>[callback.toRawHandle()]);
+    }
   }
 
   /// Promote the geofencing service to a foreground service.
@@ -142,7 +147,7 @@ class GeofencingManager {
       throw UnsupportedError("iOS does not support 'GeofenceEvent.dwell'");
     }
     final List<dynamic> args = <dynamic>[
-      PluginUtilities.getCallbackHandle(callback).toRawHandle()
+      PluginUtilities.getCallbackHandle(callback)!.toRawHandle()
     ];
     args.addAll(region._toArgs());
     await _channel.invokeMethod('GeofencingPlugin.registerGeofence', args);
